+++ mjolnir/core/GJFNVTLangevinIntegrator.hpp
@@ -27,6 +27,7 @@ class GJFNVTLangevinIntegrator
     using forcefield_type = std::unique_ptr<ForceFieldBase<traitsT>>;
     using rng_type        = RandomNumberGenerator<traits_type>;
     using remover_type    = SystemMotionRemover<traits_type>;
+    using variable_key_type = typename system_type::variable_key_type;
 
   public:
 
@@ -36,7 +37,6 @@ class GJFNVTLangevinIntegrator
           betas_(alphas_.size()),
           bs_(alphas_.size()),
           vel_coefs_(alphas_.size()),
-          noise_(alphas_.size()),
           remover_(std::move(remover))
     {}
     ~GJFNVTLangevinIntegrator() = default;
@@ -63,6 +63,11 @@ class GJFNVTLangevinIntegrator
             {
                 sys.force(i) = math::make_coordinate<coordinate_type>(0, 0, 0);
             }
+            for(auto& kv : sys.variables())
+            {
+                auto& var = kv.second;
+                var.update(var.x(), var.v(), real_type(0));
+            }
             sys.virial() = matrix33_type(0,0,0, 0,0,0, 0,0,0);
             ff->calc_force(sys);
         }
@@ -70,7 +75,7 @@ class GJFNVTLangevinIntegrator
     }
 
     real_type step(const real_type time, system_type& sys, forcefield_type& ff,
-                   rng_type& rng);
+                   rng_type&);
 
     void update(const system_type& sys)
     {
@@ -92,11 +97,9 @@ class GJFNVTLangevinIntegrator
 
     void reset_parameters(const system_type& sys) noexcept
     {
-        alphas_   .resize(sys.size());
         betas_    .resize(sys.size());
         bs_       .resize(sys.size());
         vel_coefs_.resize(sys.size());
-        noise_    .resize(sys.size());
         const auto kBT = physics::constants<real_type>::kB() * this->temperature_;
         for(std::size_t i=0; i<sys.size(); ++i)
         {
@@ -105,6 +108,19 @@ class GJFNVTLangevinIntegrator
             this->bs_.at(i)    = 1.0 / (1.0 + alpha * dt_ * 0.5 / m);
             this->betas_.at(i) = std::sqrt(2 * alpha * kBT * dt_);
             this->vel_coefs_.at(i) = 1.0 - alpha * bs_.at(i) * dt_ / m;
+        }
+
+        for(const auto& kv : sys.variables())
+        {
+            const auto& key = kv.first;
+            const auto& var = kv.second;
+
+            dynvar_params_type param;
+            param.alpha    = var.m() * var.gamma();
+            param.beta     = std::sqrt(2 * param.alpha * kBT * dt_);
+            param.b        = real_type(1) / (real_type(1) + param.alpha * dt_ / (2 * var.m()));
+            param.vel_coef = real_type(1) - param.alpha * param.b * dt_ / var.m();
+            params_for_dynvar_[key] = param;
         }
         return;
     }
@@ -126,9 +142,17 @@ class GJFNVTLangevinIntegrator
     std::vector<real_type> betas_;
     std::vector<real_type> bs_;
     std::vector<real_type> vel_coefs_;
-    std::vector<coordinate_type> noise_;
 
     remover_type remover_;
+
+    struct dynvar_params_type
+    {
+        real_type alpha; // m * gamma
+        real_type beta;
+        real_type b;
+        real_type vel_coef;
+    };
+    std::map<variable_key_type, dynvar_params_type> params_for_dynvar_;
 };
 
 template<typename traitsT>
@@ -139,34 +163,48 @@ GJFNVTLangevinIntegrator<traitsT>::step(const real_type time,
     real_type largest_disp2(0);
     for(std::size_t i=0; i<sys.size(); ++i)
     {
-        const auto rm = sys.rmass(i);  // reciprocal mass
+        const auto rm  = sys.rmass(i);        // reciprocal mass
+        const auto r2m = real_type(0.5) * rm; // 1 / 2m
+
         auto&       p = sys.position(i);
         auto&       v = sys.velocity(i);
         auto&       f = sys.force(i);
-        auto&    beta = this->noise_[i];
         const auto& b = this->bs_[i];
+        const auto& a = this->vel_coefs_[i];
 
-        // beta^(n+1) = N(0, sqrt(2alpha kBT dt))
-        // v^(n+1/2)  = v^(n) + dt/2m f^(n) + beta^(n+1) / 2m
-        // r^(n+1)    = r^(n) + b dt v^(n+1/2)
-        // f^(n+1)    = -dU/dr|r^(n+1)
-        // v^(n+1)    = (1 - alpha*b*dt/m) * v^(n+1/2) + dt/2m f^(n+1) + beta^(n+1) / 2m
+        // r_n+1 =  r_n + bdtv_n + bdt^2/2m f_n + bdt/2m beta_n+1
+        // v_n+1 = av_n + dt/2m (af_n + f_n+1) + b/m beta_n+1
 
-        beta = this->gen_R(rng) * betas_[i] * rm;
-        v   += ((dt_ * rm) * f + beta) * real_type(0.5);
+        const auto beta = this->gen_R(rng) * betas_[i]; // beta_n+1
 
-        const auto dp = (b * dt_) * v;
+        const auto dp = (b * dt_) * (v + r2m * (dt_ * f + beta));
 
-        p   += dp;
-        p    = sys.adjust_position(p);
-        v   *= vel_coefs_[i];
+        p = sys.adjust_position(p + dp);
+        v = a * (v + (dt_ * r2m) * f) + b * rm * beta;
 
-        // reset force
-        f = math::make_coordinate<coordinate_type>(0, 0, 0);
+        f = math::make_coordinate<coordinate_type>(0, 0, 0); // reset force
 
         // collect largest displacement
         largest_disp2 = std::max(largest_disp2, math::length_sq(dp));
     }
+    for(auto& kv : sys.variables()) // assume there are only a few dynvars
+    {
+        const auto& key = kv.first;
+        auto& param = params_for_dynvar_.at(key);
+        auto& var = kv.second;
+
+        const auto beta = rng.gaussian() * param.beta;
+
+        const auto next_x = var.x() + (param.b * dt_) * (var.v() +
+            (real_type(0.5) / var.m()) * (dt_ * var.f() + beta));
+
+        const auto next_v = param.vel_coef * (var.v() +
+                (real_type(0.5) * dt_ / var.m()) * var.f()) +
+                (param.b / var.m()) * beta;
+
+        var.update(next_x, next_v, real_type(0));
+    }
+
     sys.virial() = matrix33_type(0,0,0, 0,0,0, 0,0,0);
 
     // update neighbor list; reduce margin, reconstruct the list if needed
@@ -178,9 +216,17 @@ GJFNVTLangevinIntegrator<traitsT>::step(const real_type time,
         const auto  rm   = sys.rmass(i);  // reciprocal math
         const auto& f    = sys.force(i);
         auto&       v    = sys.velocity(i);
-        const auto& beta = this->noise_[i];
 
-        v += ((dt_ * rm) * f + beta) * real_type(0.5);
+        v += (dt_ * real_type(0.5) * rm) * f;
+    }
+    for(auto& kv : sys.variables()) // assume there are only a few dynvars
+    {
+        auto& var = kv.second;
+
+        const auto next_v = var.v() +
+            (dt_ * real_type(0.5) / var.m()) * var.f();
+
+        var.update(var.x(), next_v, var.f());
     }
 
     // remove net rotation/translation