+++ mjolnir/core/gBAOABLangevinIntegrator.hpp
@@ -26,6 +26,7 @@ class gBAOABLangevinIntegrator
     using forcefield_type = std::unique_ptr<ForceFieldBase<traitsT>>;
     using rng_type         = RandomNumberGenerator<traits_type>;
     using remover_type     = SystemMotionRemover<traits_type>;
+    using variable_key_type = typename system_type::variable_key_type;
 
   public:
 
@@ -72,6 +73,20 @@ class gBAOABLangevinIntegrator
             this->noise_coeff_ .at(i) = std::sqrt(
                     kBT * (1 - std::exp(2 * gamma_dt)) * sys.rmass(i));
         }
+
+        for(const auto& kv : sys.variables())
+        {
+            const auto& key = kv.first;
+            const auto& var = kv.second;
+
+            // force is not initialized yet
+            dynvar_params param;
+            param.exp_gamma_dt = std::exp(-var.gamma() * this->dt_);
+            param.noise_coeff  = std::sqrt(kBT *
+                    (real_type(1) - std::exp(-2 * var.gamma() * this->dt_)) /
+                    var.m());
+            params_for_dynvar_[key] = param;
+        }
         return;
     };
 
@@ -88,8 +103,10 @@ class gBAOABLangevinIntegrator
         const auto& constraint_ff = ff->constraint();
         const auto& constraints   = constraint_ff.constraints();
 
+        const auto max_iteration = constraint_ff.max_iteration();
+
         std::size_t rattle_step = 0;
-        while(rattle_step < constraint_ff.max_iteration())
+        while(rattle_step < max_iteration)
         {
             bool corrected = false;
             for(std::size_t i=0; i<constraints.size(); ++i)
@@ -132,7 +149,7 @@ class gBAOABLangevinIntegrator
             ++rattle_step;
         }
 
-        if(constraint_ff.max_iteration() <= rattle_step)
+        if(max_iteration != 0 && max_iteration <= rattle_step)
         {
             MJOLNIR_GET_DEFAULT_LOGGER();
             MJOLNIR_LOG_FUNCTION();
@@ -147,8 +164,10 @@ class gBAOABLangevinIntegrator
         const auto& constraint_ff = ff->constraint();
         const auto& constraints   = constraint_ff.constraints();
 
+        const auto max_iteration = constraint_ff.max_iteration();
+
         std::size_t rattle_step = 0;
-        while(rattle_step < constraint_ff.max_iteration())
+        while(rattle_step < max_iteration)
         {
             bool corrected = false;
             for(std::size_t i=0; i<constraints.size(); ++i)
@@ -181,7 +200,7 @@ class gBAOABLangevinIntegrator
             ++rattle_step;
         }
 
-        if(constraint_ff.max_iteration() <= rattle_step)
+        if(max_iteration != 0 && max_iteration <= rattle_step)
         {
             MJOLNIR_GET_DEFAULT_LOGGER();
             MJOLNIR_LOG_FUNCTION();
@@ -213,6 +232,12 @@ class gBAOABLangevinIntegrator
     std::vector<coordinate_type> old_position_;
     std::vector<coordinate_type> old_pos_rattle_;
 
+    struct dynvar_params
+    {
+        real_type exp_gamma_dt;
+        real_type noise_coeff;
+    };
+    std::map<variable_key_type, dynvar_params> params_for_dynvar_;
 };
 
 template<typename traitsT>
@@ -244,6 +269,11 @@ void gBAOABLangevinIntegrator<traitsT>::initialize(
             system.force(i) = math::make_coordinate<coordinate_type>(0, 0, 0);
         }
         system.virial() = matrix33_type(0,0,0, 0,0,0, 0,0,0);
+        for(auto& kv : system.variables())
+        {
+            auto& var = kv.second;
+            var.update(var.x(), var.v(), real_type(0));
+        }
         ff->calc_force(system);
     }
 
@@ -284,6 +314,12 @@ gBAOABLangevinIntegrator<traitsT>::step(
     }
     correct_velocity(sys, ff);
 
+    for(auto& kv : sys.variables())
+    {
+        auto& var = kv.second;
+        var.update(var.x(), var.v() + halfdt_ * var.f() / var.m(), var.f());
+    }
+
     // A step
     for(std::size_t correction_step=0; correction_step<correction_iter_num_; ++correction_step)
     {
@@ -294,6 +330,11 @@ gBAOABLangevinIntegrator<traitsT>::step(
         }
         correct_coordinate(sys, ff);
         correct_velocity(sys, ff);
+    }
+    for(auto& kv : sys.variables())
+    {
+        auto& var = kv.second;
+        var.update(var.x() + halfdt_ * var.v(), var.v(), var.f());
     }
 
     // O step
@@ -304,6 +345,16 @@ gBAOABLangevinIntegrator<traitsT>::step(
     }
     correct_velocity(sys, ff);
 
+    for(auto& kv : sys.variables())
+    {
+        const auto& param = params_for_dynvar_.at(kv.first);
+        auto& var = kv.second;
+
+        const real_type next_v = var.v() * param.exp_gamma_dt +
+            param.noise_coeff * rng.gaussian();
+        var.update(var.x(), next_v, var.f());
+    }
+
     // A step
     for(std::size_t correction_step=0; correction_step<correction_iter_num_; ++correction_step)
     {
@@ -315,6 +366,12 @@ gBAOABLangevinIntegrator<traitsT>::step(
         correct_coordinate(sys, ff);
         correct_velocity(sys, ff);
     }
+    for(auto& kv : sys.variables())
+    {
+        auto& var = kv.second;
+        var.update(var.x() + halfdt_ * var.v(), var.v(), var.f());
+    }
+
     // update neighbor list; reduce margin, reconstruct the list if needed;
     real_type largest_disp2(0.0);
     for(std::size_t i=0; i<sys.size(); ++i)
@@ -337,6 +394,14 @@ gBAOABLangevinIntegrator<traitsT>::step(
         sys.velocity(i) += this->halfdt_ * sys.rmass(i) * sys.force(i);
     }
     correct_velocity(sys, ff);
+
+    for(auto& kv : sys.variables())
+    {
+        auto& var = kv.second;
+        var.update(var.x(), var.v() + halfdt_ * var.f() / var.m(), var.f());
+    }
+
+    // other stuff needed
 
     remover_.remove(sys);