+++ mjolnir/input/read_forcefield.hpp
@@ -3,6 +3,7 @@
 #include <extlib/toml/toml.hpp>
 #include <mjolnir/core/ForceField.hpp>
 #include <mjolnir/forcefield/hybrid/HybridForceField.hpp>
+#include <mjolnir/forcefield/hybrid/DynamicHybridForceField.hpp>
 #include <mjolnir/forcefield/MultipleBasin/MultipleBasinForceField.hpp>
 #include <mjolnir/forcefield/MultipleBasin/MultipleBasin2BasinUnit.hpp>
 #include <mjolnir/forcefield/MultipleBasin/MultipleBasin3BasinUnit.hpp>
@@ -113,11 +114,11 @@ read_multiple_basin_forcefield(const toml::value& root, const toml::value& simul
     // forcefields.type = "MultipleBasin"
     //
     // [[simulator.forcefields.units]]
-    // basins   = ["apo1", "open1", "close1"]
-    // dVs      = [   0.0,     1.2,      6.0]
-    // delta.apo1-open1   = 6.0
-    // delta.open1-close1 = 5.0
-    // delta.close1-apo1  = 4.0
+    // basins   = ["open1", "close1"]
+    // dVs      = [    0.0,      6.0]
+    // delta    = 100.0
+    // k_chi    = 100.0
+    // chi_0    = 0.5
     //
     // [[simulator.forcefields.units]]
     // basins = ["apo2", "open2", "close2"]
@@ -262,9 +263,12 @@ read_multiple_basin_forcefield(const toml::value& root, const toml::value& simul
             auto ff1 = read_forcefield_elements(ffs.at(names.at(0)).second);
             auto ff2 = read_forcefield_elements(ffs.at(names.at(1)).second);
 
+            const auto k_chi = toml::find_or<real_type>(unit, "k_chi", real_type(0.0));
+            const auto chi_0 = toml::find_or<real_type>(unit, "chi_0", real_type(0.0));
+
             units.push_back(make_unique<MultipleBasin2BasinUnit<traitsT>>(delta,
                     names.at(0), names.at(1), dVs.at(0), dVs.at(1),
-                    std::move(ff1), std::move(ff2)));
+                    std::move(ff1), std::move(ff2), k_chi, chi_0));
         }
         else if(names.size() == 3)
         {
@@ -431,11 +435,42 @@ read_hybrid_forcefield(const toml::value& root, const toml::value& simulator)
         throw_exception<std::runtime_error>("[error] Hybrid Forcefield require 2 forcefields");
     }
 
-    return make_unique<HybridForceField<traitsT>>(
-            toml::find<real_type>(simulator, "forcefields", "lambda"),
-            read_default_forcefield<traitsT>(root, 0),
-            read_default_forcefield<traitsT>(root, 1)
-        );
+    const auto& forcefields = simulator.at("forcefields");
+    if( ! forcefields.contains("lambda"))
+    {
+        MJOLNIR_LOG_ERROR("Hybrid ForceField requires lambda value");
+        throw_exception<std::runtime_error>("[error] Hybrid Forcefield requires lambda value");
+    }
+
+    if(forcefields.at("lambda").is_floating())
+    {
+        return make_unique<HybridForceField<traitsT>>(
+                toml::find<real_type>(simulator, "forcefields", "lambda"),
+                read_default_forcefield<traitsT>(root, 0),
+                read_default_forcefield<traitsT>(root, 1)
+            );
+    }
+    else if(toml::find<std::string>(forcefields, "lambda") == "dynamic")
+    {
+        // considering umbrella sampling (could be replaced)
+        const auto k  = toml::find_or<real_type>(forcefields, "k_lambda", 0.0);
+        const auto v0 = toml::find_or<real_type>(forcefields, "v0_lambda", 0.0);
+
+        return make_unique<DynamicHybridForceField<traitsT>>(k, v0,
+                read_default_forcefield<traitsT>(root, 0),
+                read_default_forcefield<traitsT>(root, 1)
+            );
+    }
+    else
+    {
+        throw std::runtime_error(toml::format_error("mjolnir::read_hybrid_forcefield: "
+            "unknown lambda type", simulator.at("forcefields").at("lambda"),
+            "here", {"expected one of the following: ",
+                "- (real type): a constant lambda works as `lambda * V1 + (1 - lambda) * V2`.",
+                "- \"dynamic\": lambda also moves as a dynamic variable."
+            }));
+
+    }
 }
 
 template<typename traitsT>