+++ mjolnir/omp/System.hpp
@@ -32,11 +32,15 @@ class System<OpenMPSimulatorTraits<realT, boundaryT>>
     template<typename T>
     using cache_aligned_allocator = aligned_allocator<T, cache_alignment>;
 
+    using dynamic_variable_type = DynamicVariable<real_type>;
+    using variable_key_type = std::string;
+    using variables_type = std::map<variable_key_type, dynamic_variable_type>;
+
   public:
 
     System(const std::size_t num_particles, const boundary_type& bound)
         : velocity_initialized_(false), force_initialized_(false),
-          boundary_(bound), attributes_(),
+          boundary_(bound), attributes_{}, variables_{},
           virial_(0,0,0, 0,0,0, 0,0,0),
           virial_threads_(omp_get_max_threads(),
                           matrix33_type(0,0,0, 0,0,0, 0,0,0)),
@@ -84,6 +88,16 @@ class System<OpenMPSimulatorTraits<realT, boundaryT>>
             math::X(this->velocity(i)) = rng.gaussian(0, vel_coef);
             math::Y(this->velocity(i)) = rng.gaussian(0, vel_coef);
             math::Z(this->velocity(i)) = rng.gaussian(0, vel_coef);
+        }
+
+        // generate random force for dynamic variables
+        for(auto& kv : this->variables_)
+        {
+            auto& var = kv.second;
+            if( ! is_finite(var.v())) // not initialized
+            {
+                var.update(var.x(), rng.gaussian(0, std::sqrt(kBT / var.m())), var.f());
+            }
         }
         MJOLNIR_LOG_NOTICE("done.");
         return;
@@ -192,6 +206,13 @@ class System<OpenMPSimulatorTraits<realT, boundaryT>>
     bool   has_attribute(const std::string& key) const {return attributes_.count(key) == 1;}
     attribute_type const& attributes() const noexcept {return attributes_;}
 
+    // dynamic variables in a system.
+    dynamic_variable_type const& variable(const std::string& key) const {return variables_.at(key);}
+    dynamic_variable_type&       variable(const std::string& key)       {return variables_[key];}
+    bool          has_variable(const std::string& key) const {return variables_.count(key) == 1;}
+    variables_type const& variables() const noexcept {return variables_;}
+    variables_type&       variables()       noexcept {return variables_;}
+
     bool  velocity_initialized() const noexcept {return velocity_initialized_;}
     bool& velocity_initialized()       noexcept {return velocity_initialized_;}
 
@@ -207,6 +228,7 @@ class System<OpenMPSimulatorTraits<realT, boundaryT>>
     bool           velocity_initialized_, force_initialized_;
     boundary_type  boundary_;
     attribute_type attributes_;
+    variables_type variables_;
 
     matrix33_type  virial_;
     std::vector<matrix33_type, cache_aligned_allocator<matrix33_type>> virial_threads_;